+++ packages/panels/src/components/nav-bar/NavBarButton.tsx
@@ -3,9 +3,9 @@ import * as React from "react";
 import styles from "./NavBarButton.module.css";
 import cx from "classnames";
 
-export type NavBarButtonProps = Extract<FlatButtonProps, { as?: "button" }> & {
+export interface NavBarButtonProps extends FlatButtonProps {
   selected?: boolean;
-};
+}
 
 export const NavBarButton: React.FC<NavBarButtonProps> = ({
   selected,