+++ packages/tooltip/src/components/button-with-popover/ButtonWithPopover.tsx
@@ -2,15 +2,12 @@ import { FlatButton, FlatButtonProps } from "@stenajs-webui/elements";
 import * as React from "react";
 import { Popover, PopoverProps } from "../popover/Popover";
 
-export type ButtonWithPopoverProps = Extract<
-  FlatButtonProps,
-  { as?: "button" }
-> & {
+export interface ButtonWithPopoverProps extends FlatButtonProps {
   children?: PopoverProps["children"];
   visible?: boolean;
   onShow?: () => void;
   onHide?: () => void;
-};
+}
 
 export const ButtonWithPopover: React.FC<ButtonWithPopoverProps> = ({
   children,